Updating tests and removing unneeded files
